CheckStyle and JavaDocs
