+++ src/main/java/seedu/address/logic/commands/contactlistcommands/AddContactCommand.java
@@ -1,10 +1,13 @@
-package seedu.address.logic.commands.contactListCommands;
+package seedu.address.logic.commands.contactlistcommands;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 
+/**
+ * Adds a contact to the contact list.
+ */
 public class AddContactCommand extends Command {
     public static final String COMMAND_WORD = "addcontact";
