+++ src/main/java/seedu/address/logic/commands/contactlistcommands/DeleteContactCommand.java
@@ -1,12 +1,15 @@
-package seedu.address.logic.commands.contactListCommands;
+package seedu.address.logic.commands.contactlistcommands;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 
-public class HelpContactCommand extends Command {
-    public static final String COMMAND_WORD = "helpcontact";
+/**
+ * Deletes a specific contact from the contact list.
+ */
+public class DeleteContactCommand extends Command {
+    public static final String COMMAND_WORD = "deletecontact";
 
     @Override
     public CommandResult execute(Model model) throws CommandException {
