+++ src/main/java/seedu/address/logic/commands/contactlistcommands/HelpContactCommand.java
@@ -1,12 +1,15 @@
-package seedu.address.logic.commands.contactListCommands;
+package seedu.address.logic.commands.contactlistcommands;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 
-public class DeleteContactCommand extends Command {
-    public static final String COMMAND_WORD = "deletecontact";
+/**
+ * Format full help instructions for every contact related command for display.
+ */
+public class HelpContactCommand extends Command {
+    public static final String COMMAND_WORD = "helpcontact";
 
     @Override
     public CommandResult execute(Model model) throws CommandException {
