+++ src/main/java/seedu/address/logic/commands/contactlistcommands/ListContactCommand.java
@@ -1,12 +1,15 @@
-package seedu.address.logic.commands.contactListCommands;
+package seedu.address.logic.commands.contactlistcommands;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 
-public class ViewContactCommand extends Command {
-    public static final String COMMAND_WORD = "viewcontact";
+/**
+ * Lists all the existing contacts in the list to display.
+ */
+public class ListContactCommand extends Command {
+    public static final String COMMAND_WORD = "listcontact";
 
     @Override
     public CommandResult execute(Model model) throws CommandException {
