+++ src/main/java/seedu/address/logic/commands/contactlistcommands/ViewContactCommand.java
@@ -1,12 +1,15 @@
-package seedu.address.logic.commands.contactListCommands;
+package seedu.address.logic.commands.contactlistcommands;
 
 import seedu.address.logic.commands.Command;
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 
-public class ListContactCommand extends Command {
-    public static final String COMMAND_WORD = "listcontact";
+/**
+ * Displays a specific contact in the contact list.
+ */
+public class ViewContactCommand extends Command {
+    public static final String COMMAND_WORD = "viewcontact";
 
     @Override
     public CommandResult execute(Model model) throws CommandException {
