+++ src/main/java/seedu/address/logic/parser/contactlistparsers/AddContactParser.java
@@ -1,9 +1,12 @@
-package seedu.address.logic.parser.contactListParsers;
+package seedu.address.logic.parser.contactlistparsers;
 
-import seedu.address.logic.commands.contactListCommands.AddContactCommand;
+import seedu.address.logic.commands.contactlistcommands.AddContactCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 
+/**
+ * Parses input arguments and creates a new AddContactCommand object
+ */
 public class AddContactParser implements Parser<AddContactCommand> {
     @Override
     public AddContactCommand parse(String userInput) throws ParseException {
