+++ src/main/java/seedu/address/logic/parser/contactlistparsers/DeleteContactParser.java
@@ -1,9 +1,12 @@
-package seedu.address.logic.parser.contactListParsers;
+package seedu.address.logic.parser.contactlistparsers;
 
-import seedu.address.logic.commands.contactListCommands.DeleteContactCommand;
+import seedu.address.logic.commands.contactlistcommands.DeleteContactCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 
+/**
+ * Parses input arguments and creates a new DeleteContactCommand object
+ */
 public class DeleteContactParser implements Parser<DeleteContactCommand> {
     @Override
     public DeleteContactCommand parse(String userInput) throws ParseException {
