+++ src/main/java/seedu/address/logic/parser/contactlistparsers/EditContactParser.java
@@ -1,9 +1,12 @@
-package seedu.address.logic.parser.contactListParsers;
+package seedu.address.logic.parser.contactlistparsers;
 
-import seedu.address.logic.commands.contactListCommands.EditContactCommand;
+import seedu.address.logic.commands.contactlistcommands.EditContactCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 
+/**
+ * Parses input arguments and creates a new EditContactCommand object
+ */
 public class EditContactParser implements Parser<EditContactCommand> {
     @Override
     public EditContactCommand parse(String userInput) throws ParseException {
