+++ src/main/java/seedu/address/logic/parser/contactlistparsers/FindContactParser.java
@@ -1,9 +1,12 @@
-package seedu.address.logic.parser.contactListParsers;
+package seedu.address.logic.parser.contactlistparsers;
 
-import seedu.address.logic.commands.contactListCommands.FindContactCommand;
+import seedu.address.logic.commands.contactlistcommands.FindContactCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 
+/**
+ * Parses input arguments and creates a new FindContactCommand object
+ */
 public class FindContactParser implements Parser<FindContactCommand> {
     @Override
     public FindContactCommand parse(String userInput) throws ParseException {
