+++ src/main/java/seedu/address/logic/parser/contactlistparsers/ViewContactParser.java
@@ -1,9 +1,12 @@
-package seedu.address.logic.parser.contactListParsers;
+package seedu.address.logic.parser.contactlistparsers;
 
-import seedu.address.logic.commands.contactListCommands.ViewContactCommand;
+import seedu.address.logic.commands.contactlistcommands.ViewContactCommand;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.exceptions.ParseException;
 
+/**
+ * Parses input arguments and creates a new ViewContactCommand object
+ */
 public class ViewContactParser implements Parser<ViewContactCommand> {
     @Override
     public ViewContactCommand parse(String userInput) throws ParseException {
